Bug fixes for MongoAlchemyDbAdapter.

+++ flask_user/tests/conftest.py
@@ -33,7 +33,8 @@ def app(request):
 def db(app, request):
     """Session-wide test database."""
     def teardown():
-        app.db.drop_all()
+        if hasattr(app.db, 'drop_all'):
+            app.db.drop_all()
 
     if hasattr(app.db, 'create_all'):
         app.db.create_all()
